isolated passwordViewable to PasswordBox component
